Reset folder structure
