+++ ui/student/check-in/YouthDetails.tsx
@@ -1,8 +1,8 @@
 
 import { useState } from 'react';
-import styles from '../CheckIn.module.scss'
-import { StepProps } from '../page';
+import styles from '@/app/(protected)/student/check-in/CheckIn.module.scss'
 import ProgressBar from './ProgressBar';
+import { StepProps } from '@/app/(protected)/student/check-in/page';
 
 interface YouthDetailsProps extends StepProps {
     heading: string;
